leo-worker/archive-2025-12-17/archive/apps/ai-lawyer-sprints1/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.25
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0
types-passlib==1.7.7.20240106
types-python-jose==3.3.4.20240106
PyPDF2==3.0.1
jinja2==3.1.3
reportlab==4.0.7
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0
# Email functionality is handled via API, not nodemailer